working copy to fix cart

+++ ExtremeBicycle/Models/Entities/OrderDetails.cs
@@ -11,7 +11,7 @@ namespace ExtremeBicycle.Models.Entities {
 
         public int ProductID { get; set; }
 
-        public double? UnitPrice { get; set; }
+        public decimal? UnitPrice { get; set; }
 
         public int? Quantity { get; set; }
